rename file, take some of matt's suggestions
